+++ packages/teleport/__tests__/exporter.spec.js
@@ -1,9 +1,24 @@
 // @ts-check
+import path from 'node:path';
+import {fileURLToPath} from 'node:url';
 import {expect} from 'chai';
 import nock from 'nock';
 import {exporter} from '../lib/api.js';
+import {exportUserRows} from '../lib/exporter.js';
+import {createKnexProxy} from '../lib/shared/db.js';
+import {rawUserExport} from './fixtures/user-export.js';
+import {runSqlFile, useDatabase} from './_utils/test-db.js';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+
+const knex = useDatabase();
 
 describe('Unit > Exporter', function () {
+	before(async function () {
+		await runSqlFile(knex, path.resolve(__dirname, '../fixtures/database-ddl.sql'));
+		await runSqlFile(knex, path.resolve(__dirname, '../fixtures/database-dml.sql'));
+	});
+
 	it('getExport wraps request to server endpoint', async function () {
 		try {
 			nock('https://gradebook.app')
@@ -28,5 +43,24 @@ describe('Unit > Exporter', function () {
 		} finally {
 			nock.cleanAll();
 		}
+	});
+
+	it('exportUserRows can handle users that do not exist', async function () {
+		const proxy = createKnexProxy(knex, null);
+		expect(await exportUserRows(proxy, '012345678901234567890123')).to.deep.equal({
+			error: 'Unable to find user',
+		});
+	});
+
+	it('exportUserRows generates an export', async function () {
+		const proxy = createKnexProxy(knex, null);
+		const userExport = await exportUserRows(proxy, '6158081ca3c0c1619d74088a');
+
+		if ('error' in userExport) { // Only needed for type checking
+			expect(userExport.error).to.not.be.ok;
+			return;
+		}
+
+		expect(userExport).to.deep.equal(rawUserExport);
 	});
 });
